email_Feature/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
resend>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0
